pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "palantir-ontology-agents"
version = "0.1.0"
description = "Multi-agent ontology workflow system demonstrating AIP-style agentic intelligence analysis with LangGraph"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Hashwanth Sutharapu"}
]
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-community>=0.3.0",
    "tavily-python>=0.5.0",
    "streamlit>=1.40.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
